Certificates Controller

+++ CertificatesManager/CertificatesManager/Controllers/CertificatesController.cs
@@ -12,6 +12,7 @@ using Microsoft.AspNet.Identity.Owin;
 
 namespace CertificatesManager.Controllers
 {
+    [Authorize]
     public class CertificatesController : Controller
     {
         private ApplicationDbContext db = new ApplicationDbContext();
@@ -21,17 +22,23 @@ namespace CertificatesManager.Controllers
         {
             ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(System.Web.HttpContext.Current.User.Identity.GetUserId());
 
-            return View(db.Certificates.ToList());
+            return View(db.Certificates
+                    .Where(x => x.EOSAuthorAccount == user.EOSAccountName || x.EOSOwnerAccount == user.EOSAccountName)
+                    .ToList()
+                );
         }
 
         // GET: Certificates/Details/5
         public ActionResult Details(int? id)
         {
+            ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(System.Web.HttpContext.Current.User.Identity.GetUserId());
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Certificate certificate = db.Certificates.Find(id);
+            Certificate certificate = db.Certificates.Where(x => x.EOSAuthorAccount == user.EOSAccountName || x.EOSOwnerAccount == user.EOSAccountName).SingleOrDefault();
+
             if (certificate == null)
             {
                 return HttpNotFound();
@@ -96,11 +103,14 @@ namespace CertificatesManager.Controllers
         // GET: Certificates/Delete/5
         public ActionResult Delete(int? id)
         {
+            ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(System.Web.HttpContext.Current.User.Identity.GetUserId());
+
             if (id == null)
             {
                 return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
             }
-            Certificate certificate = db.Certificates.Find(id);
+            Certificate certificate = db.Certificates.Where(x => x.EOSAuthorAccount == user.EOSAccountName || x.EOSOwnerAccount == user.EOSAccountName).SingleOrDefault();
+
             if (certificate == null)
             {
                 return HttpNotFound();
@@ -113,7 +123,13 @@ namespace CertificatesManager.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult DeleteConfirmed(int id)
         {
-            Certificate certificate = db.Certificates.Find(id);
+            ApplicationUser user = System.Web.HttpContext.Current.GetOwinContext().GetUserManager<ApplicationUserManager>().FindById(System.Web.HttpContext.Current.User.Identity.GetUserId());
+
+            Certificate certificate = db.Certificates.Where(x => x.EOSAuthorAccount == user.EOSAccountName || x.EOSOwnerAccount == user.EOSAccountName).SingleOrDefault();
+            if (certificate == null)
+            {
+                return HttpNotFound();
+            }
             db.Certificates.Remove(certificate);
             db.SaveChanges();
             return RedirectToAction("Index");
